pyproject.toml
[project]
name = "course-dev-ide-with-agents"
version = "0.1.0"
description = "Collaborative intelligence infrastructure for course content production"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Juanl" }]
keywords = ["edtech", "course-production", "multi-agent", "content-creation"]

dependencies = [
    "requests>=2.31,<3",
    "python-dotenv>=1.0,<2",
    "PyJWT>=2.8,<3",
    "pydantic>=2.7,<3",
    "aiohttp>=3.9,<4",
    "pyyaml>=6.0,<7",
    "imageio-ffmpeg>=0.5,<1",
    "pypdf>=5.0,<6",
    "pillow>=11.0,<12",
    "python-pptx>=1.0,<2",
    "python-docx>=1.1,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-timeout>=2.3,<3",
    "responses>=0.25,<1",
    "ruff>=0.4,<1",
    "pre-commit>=3.7,<5",
    "jsonschema>=4.0,<5",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# run_hud.py generates an HTML/CSS dashboard; embedded CSS/HTML strings in
# Python string literals are long by necessity and don't benefit from
# line-wrapping. Exempt E501 (line-too-long) for this file only.
"scripts/utilities/run_hud.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "skills/source-wrangler/scripts/tests",
    "skills/compositor/scripts/tests",
    "skills/bmad-agent-marcus/scripts/tests",
    "skills/bmad-agent-texas/scripts/tests",
    "skills/kling-video/scripts/tests",
]
pythonpath = [".", "skills/bmad-agent-texas/scripts"]  # retrieval/ package (Story 27-0)
asyncio_mode = "auto"
markers = [
    "live_api: tests that call live third-party APIs; excluded by default unless --run-live is provided",
    "live_api_e2e: slower/flakier live end-to-end API tests; excluded by default unless --run-live-e2e is provided",
    "trial_critical: dev-side tests guarding the pre-Prompt-1 trial path (run_constants validator, preflight receipt, pack-doc schema lockstep). Run before firing any trial production run via `pytest -m trial_critical`. Not a runtime gate — a dev confidence check.",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
